# Admin/bin/Debug/netcoreapp3.1/Docs/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Admin API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/devices/mine": {
      "get": {
        "tags": [
          "Devices"
        ],
        "operationId": "Devices_GetDevices",
        "parameters": [
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/SortKey3"
            },
            "x-position": 1
          },
          {
            "name": "OrderBy",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Order"
            },
            "x-position": 2
          },
          {
            "name": "DomainId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid",
              "nullable": true
            },
            "x-position": 3
          },
          {
            "name": "DeviceGroupId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid",
              "nullable": true
            },
            "x-position": 4
          },
          {
            "name": "Name",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 5
          },
          {
            "name": "ProductName",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 6
          },
          {
            "name": "LteModuleId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid",
              "nullable": true
            },
            "x-position": 7
          },
          {
            "name": "StartDateFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "x-position": 8
          },
          {
            "name": "StartDateTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "x-position": 9
          },
          {
            "name": "EndDateFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "x-position": 10
          },
          {
            "name": "EndDateTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "x-position": 11
          },
          {
            "name": "IncludingEndDateNotSet",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "x-position": 12
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 2147483647.0,
              "minimum": 1.0,
              "nullable": true
            },
            "x-position": 13
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 1000.0,
              "minimum": 1.0
            },
            "x-position": 14
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseOfDeviceDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "UserAdmin"
            ]
          }
        ]
      }
    },
    "/api/devices": {
      "get": {
        "tags": [
          "Devices"
        ],
        "operationId": "Devices_GetDevices2",
        "parameters": [
          {
            "name": "OrganizationCode",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/SortKey3"
            },
            "x-position": 2
          },
          {
            "name": "OrderBy",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Order"
            },
            "x-position": 3
          },
          {
            "name": "DomainId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid",
              "nullable": true
            },
            "x-position": 4
          },
          {
            "name": "DeviceGroupId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid",
              "nullable": true
            },
            "x-position": 5
          },
          {
            "name": "Name",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 6
          },
          {
            "name": "ProductName",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 7
          },
          {
            "name": "LteModuleId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid",
              "nullable": true
            },
            "x-position": 8
          },
          {
            "name": "StartDateFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "x-position": 9
          },
          {
            "name": "StartDateTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "x-position": 10
          },
          {
            "name": "EndDateFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "x-position": 11
          },
          {
            "name": "EndDateTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "x-position": 12
          },
          {
            "name": "IncludingEndDateNotSet",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "x-position": 13
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 2147483647.0,
              "minimum": 1.0,
              "nullable": true
            },
            "x-position": 14
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 1000.0,
              "minimum": 1.0
            },
            "x-position": 15
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseOfDeviceDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "SuperAdmin"
            ]
          }
        ]
      }
    },
    "/api/devices/mine/{id}": {
      "get": {
        "tags": [
          "Devices"
        ],
        "operationId": "Devices_GetDevice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "UserAdmin"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "Devices"
        ],
        "operationId": "Devices_PutDeviceMine",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "param",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutDeviceParam"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "UserAdmin"
            ]
          }
        ]
      }
    },
...
    "/api/devices/mine/csv": {
      "get": {
        "tags": [
          "Devices"
        ],
        "operationId": "Devices_GetDevicesCsv",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "UserAdmin"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "Devices"
        ],
        "operationId": "Devices_PostDevicesCsv",
        "parameters": [
          {
            "type": "file",
            "name": "csv",
            "in": "formData",
            "schema": {
              "type": "string",
              "format": "binary",
              "nullable": true
            },
            "nullable": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Device"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "UserAdmin"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Order": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Asc",
          "Desc"
        ],
        "enum": [
          0,
          1
        ]
      },
...
      "PaginatedResponseOfDeviceDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "results": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/DeviceDto"
            }
          }
        }
      },
...
      "SortKey3": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "DomainName",
          "Name",
          "ProductName",
          "LteModuleName",
          "StartDate",
          "EndDate"
        ],
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5
        ]
      },
      "PutDeviceParam": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "useTpm",
          "windowsSignInListCacheDays"
        ],
        "properties": {
          "managedNumber": {
            "type": "string",
            "nullable": true
          },
          "serialNumber": {
            "type": "string",
            "nullable": true
          },
          "productName": {
            "type": "string",
            "nullable": true
          },
          "lteModuleId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "organizationClientAppId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "useTpm": {
            "type": "boolean"
          },
          "windowsSignInListCacheDays": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    },
    "securitySchemes": {
      "JWT": {
        "type": "apiKey",
        "description": "Type into the textbox: Bearer {your JWT token}.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "JWT": []
    }
  ],
...
}